Added pseudo-automated batch build file
It allows to keep the source files outside the Forge/mcp/src folder,
which is useful for developing and versioning of several mods. It
requires that there is FORGE_HOME environment variable, and if WinRAR is
installed in the default location, it will perform the archiving into
the final file as well.